requirements.txt
numpy==1.21.4
Pillow==8.4.0
PyQt5==5.15.6
opencv-python==4.5.3.56